Improve skeleton file logic
The `++edit` option was for `read`, which causes additional buffer
options to be automatically set based on the `filetype` for the file
being loaded. Additionally, the first empty line that pops up after the
`:read` is deleted and the `modified` option for the current buffer is
set to `false` so that the buffer can be immediately closed without
being asked if we want to discard the unsaved changes.